[refactor] 03/02 함수 화살표 함수로 변경 #192
