+++ application/usecases/novel/dto/NovelEpisode.ts
